PROJECTS/bingx-connector-v2/requirements.txt
requests>=2.28.0
pyyaml>=6.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
